feat: Add ability to add Version manually
User Story:
As a developer I want to be able to define the Version manually.

Background:
If for example you use CLI on CI level, you probably use Semantic Release (https://www.npmjs.com/package/semantic-release) to define the Versions of the Application. In this case, the versions created by react-native-cli-bump-version and Semantic Release will be different. To be able to use it with SR it will be good to add additional parameter to this CLI